feat: migrate auth pages
