+++ src/app/(auth)/signup/page.tsx
@@ -1,5 +1,30 @@
-import { SignupForm } from "../components/SignupForm"
+"use client"
+import { BlitzPage } from "@blitzjs/next"
+import Link from "next/link"
+import { useRouter } from "next/compat/router"
+import { SignupForm } from "src/auth/components/SignupForm"
+import Layout from "src/core/layouts/Layout"
 
-export default function SignUpPage() {
-  return <SignupForm />
+const CreateAccountPage: BlitzPage = () => {
+  const router = useRouter()
+
+  return (
+    <Layout title="Create Account">
+      <div className="relative min-h-screen">
+        <nav className="border-ladderly-light-purple flex border bg-ladderly-off-white px-4 py-1 text-ladderly-violet-700">
+          <Link
+            href="/"
+            className="ml-auto text-gray-800 hover:text-ladderly-pink"
+          >
+            Back to Home
+          </Link>
+        </nav>
+        <div className="flex min-h-[calc(100vh-4rem)] items-center justify-center">
+          <SignupForm onSuccess={() => router && router.push("/")} />
+        </div>
+      </div>
+    </Layout>
+  )
 }
+
+export default CreateAccountPage
